Check if the file was really added to archive, and is ready for deletion

+++ apt_offline_core/AptOfflineCoreLib.py
@@ -214,7 +214,8 @@ class FetchBugReports( AptOfflineLib.Archiver ):
                 
         def AddToArchive(self, ArchiveFile, fileName):
                 if self.compress_the_file(ArchiveFile, fileName):
-                        os.unlink(fileName)
+                        if self.file_possibly_deleted is not True:
+                                os.unlink(fileName)
                 return True
         
 class ExecCmd:
